Components refactoring and redux setup finalization
