+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "project-work-v2",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "private": true,
   "dependencies": {
     "@reduxjs/toolkit": "^1.8.6",
@@ -8,18 +8,24 @@
     "@testing-library/react": "^13.0.1",
     "@testing-library/user-event": "^14.1.1",
     "@types/axios": "^0.14.0",
+    "@types/history": "^4.7.10",
     "@types/jest": "^27.4.1",
     "@types/node": "^17.0.25",
     "@types/react": "^18.0.21",
     "@types/react-dom": "^18.0.2",
     "@types/react-redux": "^7.1.24",
     "@types/react-router-dom": "^5.3.3",
+    "@types/redux-saga": "^0.10.5",
     "antd": "^4.23.6",
+    "axios": "^1.1.3",
+    "history": "4.10.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-redux": "^8.0.1",
+    "react-redux": "^8.0.4",
     "react-router-dom": "5.3.4",
     "react-scripts": "5.0.1",
+    "redux": "^4.2.0",
+    "redux-saga": "^1.2.1",
     "typescript": "^4.6.0",
     "web-vitals": "^2.1.0"
   },
@@ -46,5 +52,11 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@types/redux-logger": "^3.0.9",
+    "@types/redux-persist": "^4.3.1",
+    "redux-logger": "^3.0.6",
+    "redux-persist": "^6.0.0"
   }
 }
