+++ src/App.test.tsx
@@ -1,10 +1,12 @@
-import React from 'react';
-import { render } from '@testing-library/react';
-import { Provider } from 'react-redux';
-import { store } from './app/store';
-import App from './App';
+import React from "react";
+import App from "./App";
+import { render } from "@testing-library/react";
+import { Provider } from "react-redux";
+import configureStore from "./store";
 
-test('renders learn react link', () => {
+const { persistor, store } = configureStore();
+
+test("renders learn react link", () => {
   const { getByText } = render(
     <Provider store={store}>
       <App />
