--- src/features/auth/types/LoginArguments.ts
@@ -1,4 +1,0 @@
-export interface LoginArguments {
-  email: string;
-  password: string;
-}
