+++ src/index.tsx
@@ -1,10 +1,12 @@
 import React from "react";
+import "./index.css";
+import App from "./App";
 import { createRoot } from "react-dom/client";
 import { Provider } from "react-redux";
-import { store } from "./app/store";
-import App from "./App";
-import "./index.css";
 import { BrowserRouter } from "react-router-dom";
+import configureStore from "./store";
+
+const { persistor, store } = configureStore();
 
 const container = document.getElementById("root")!;
 const root = createRoot(container);
