+++ src/store/actions/types.ts
@@ -1,10 +1,10 @@
 // loaders
-export const LOGIN_ENABLE_LOADER = 'LOGIN_ENABLE_LOADER';
-export const LOGIN_DISABLE_LOADER = 'LOGIN_DISABLE_LOADER';
+export const LOGIN_ENABLE_LOADER = "LOGIN_ENABLE_LOADER";
+export const LOGIN_DISABLE_LOADER = "LOGIN_DISABLE_LOADER";
 
 // login
-export const LOGIN_REQUEST = 'LOGIN_REQUEST';
-export const LOGIN_RESPONSE = 'LOGIN_RESPONSE';
-export const LOGIN_LOADING_ENDED = 'LOGIN_LOADING_ENDED';
-export const LOGIN_FAILED = 'LOGIN_FAILED';
-export const LOG_OUT = 'LOG_OUT';
+export const LOGIN_REQUEST = "LOGIN_REQUEST";
+export const LOGIN_RESPONSE = "LOGIN_RESPONSE";
+export const LOGIN_LOADING_ENDED = "LOGIN_LOADING_ENDED";
+export const LOGIN_FAILED = "LOGIN_FAILED";
+export const LOG_OUT = "LOG_OUT";
